+++ src/main/java/us/aaronweiss/pixalia/net/packets/MovementPacket.java
@@ -25,7 +25,7 @@ public class MovementPacket extends Packet {
 	public String hostname() {
 		if (this.packetType.is(PacketType.INBOUND)) {
 			this.ready();
-			return Utils.readString(this.buffer.readInt(), this.buffer);
+			return Utils.readString(this.buffer.readByte(), this.buffer);
 		}
 		return null;
 	}
@@ -33,7 +33,7 @@ public class MovementPacket extends Packet {
 	public Vector position() {
 		this.ready();
 		if (this.packetType.is(PacketType.INBOUND)) {
-			this.buffer.readBytes(this.buffer.readInt());
+			this.buffer.readBytes(this.buffer.readByte());
 		}
 		return Vector.fromByteBuf(2, this.buffer);
 	}
